requirements.txt
requests>=2.25.0
pandas>=1.2.0
python-dateutil>=2.8.0
python-dotenv>=0.19.0 